requirements.txt
gunicorn
dash
dash-bootstrap-components
dash-bootstrap-templates
flask-caching
whitenoise
pandas
pyaml